Traceability: Running `viash build` and `viash test` creates a `.viash_config.yaml` file by default, which contains the processed config of the component.
