+++ src/main/scala/io/viash/ViashTest.scala
@@ -33,6 +33,7 @@ import helpers.IO
 import io.viash.helpers.data_structures._
 import io.viash.helpers.MissingResourceFileException
 import io.viash.platforms.Platform
+import io.viash.config.ConfigMeta
 
 object ViashTest {
   case class TestOutput(name: String, exitValue: Int, output: String, logFile: String, duration: Long)
@@ -246,15 +247,23 @@ object ViashTest {
           dest = Some("test_executable"),
           text = funOnlyTest.resources.head.text
         )
+        val configYaml = ConfigMeta.toMetaFile(config)
 
         // assemble full resources list for test
         val funFinal = fun.copy(resources = 
-          testBash :: // the test, wrapped in a bash script
-            exe :: // the executable, wrapped with a native platform,
+          // the test, wrapped in a bash script
+          testBash ::
+            // the executable, wrapped with a native platform,
             // to be run inside of the platform of the test
-            funOnlyTest.additionalResources ::: // other resources generated by wrapping the test script
-            fun.additionalResources ::: // other resources provided in fun.resources
-            tests.filter(!_.isInstanceOf[Script]) // other resources provided in fun.tests
+            exe :: 
+            // the config file information
+            configYaml ::
+            // other resources generated by wrapping the test script
+            funOnlyTest.additionalResources :::
+            // other resources provided in fun.resources
+            fun.additionalResources :::
+            // other resources provided in fun.tests
+            tests.filter(!_.isInstanceOf[Script])
         )
 
         // write resources to dir
